moved rotator to base
